Move codestream tests
