+++ common/artifact-manager/src/test/java/com/vmware/pscoe/iac/artifact/aria/codestream/store/CsGitWebhookStoreTest.java
@@ -12,7 +12,7 @@
  * This product may include a number of subcomponents with separate copyright notices and license terms. Your use of these subcomponents is subject to the terms and conditions of the subcomponent's license, as noted in the LICENSE file.
  * #L%
  */
-package com.vmware.pscoe.iac.artifact.store.cs;
+package com.vmware.pscoe.iac.artifact.aria.codestream.store;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.times;
@@ -36,11 +36,12 @@ import org.junit.rules.TemporaryFolder;
 import org.mockito.Mockito;
 
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLMapper;
+import com.google.gson.JsonObject;
 import com.vmware.pscoe.iac.artifact.aria.codestream.configuration.ConfigurationCs;
-import com.vmware.pscoe.iac.artifact.aria.codestream.models.CustomIntegrationVersion;
 import com.vmware.pscoe.iac.artifact.aria.codestream.rest.RestClientCs;
-import com.vmware.pscoe.iac.artifact.aria.codestream.store.CsCustomIntegrationStore;
 import com.vmware.pscoe.iac.artifact.aria.codestream.store.models.CsPackageDescriptor;
 import com.vmware.pscoe.iac.artifact.helpers.AssertionsHelper;
 import com.vmware.pscoe.iac.artifact.helpers.FsMocks;
@@ -48,11 +49,11 @@ import com.vmware.pscoe.iac.artifact.model.Package;
 import com.vmware.pscoe.iac.artifact.model.PackageFactory;
 import com.vmware.pscoe.iac.artifact.model.PackageType;
 
-public class CsCustomIntegrationStoreTest {
+public class CsGitWebhookStoreTest {
 	@Rule
 	public TemporaryFolder tempFolder = new TemporaryFolder();
 
-	protected CsCustomIntegrationStore store;
+	protected CsGitWebhookStore store;
 	protected RestClientCs restClient;
 	protected Package pkg;
 	protected ConfigurationCs config;
@@ -63,16 +64,16 @@ public class CsCustomIntegrationStoreTest {
 	void init() {
 		try {
 			tempFolder.create();
-			tempFolder.newFolder("custom-integrations");
+			tempFolder.newFolder("git-webhooks");
 
 		} catch (IOException e) {
 			throw new RuntimeException("Could not create a temp folder");
 		}
 
 		fsMocks = new FsMocks(tempFolder.getRoot());
-		store = new CsCustomIntegrationStore();
+		store = new CsGitWebhookStore();
 		restClient = Mockito.mock(RestClientCs.class);
-		pkg = PackageFactory.getInstance(PackageType.CS, tempFolder.getRoot());
+		pkg = PackageFactory.getInstance(PackageType.VRANG, tempFolder.getRoot());
 		config = Mockito.mock(ConfigurationCs.class);
 		vraNgPackageDescriptor = Mockito.mock(CsPackageDescriptor.class);
 
@@ -92,104 +93,51 @@ public class CsCustomIntegrationStoreTest {
 	}
 
 	@Test
-	void testExportDesiredCi() {
+	void testExportDesiredPipeline() {
 		// GIVEN
-		List<CustomIntegrationVersion> cis = new ArrayList<>();
-		CustomIntegrationVersion ci = new CustomIntegrationVersion();
-		cis.add(ci);
-		ci.setName("testCustomIntegration");
-		ci.setStatus("DRAFT");
-
-		List<CustomIntegrationVersion> versions = new ArrayList<>();
-		CustomIntegrationVersion version = new CustomIntegrationVersion();
-		versions.add(version);
-		version.setName("testCustomIntegration");
-		version.setVersion("1");
-		version.setStatus("RELEASED");
-
-		List<String> names = Arrays.asList(new String[] { "testCustomIntegration" });
-
-		when(restClient.getCustomIntegrations()).thenReturn(cis);
-		when(restClient.getCustomIntegrationVersions(any())).thenReturn(versions);
-		when(vraNgPackageDescriptor.getCustomIntegration()).thenReturn(names);
+		List<JsonObject> pipelines = new ArrayList<>();
+		JsonObject pipeline = new JsonObject();
+		pipelines.add(pipeline);
+		pipeline.addProperty("name", "testWebhook");
+		List<String> names = Arrays.asList(new String[] { "testWebhook" });
+		when(restClient.getProjectGitWebhooks()).thenReturn(pipelines);
+		when(vraNgPackageDescriptor.getGitWebhook()).thenReturn(names);
 
 		// TEST
 		store.exportContent();
 
-		String[] expectedPipelinefile = { "testCustomIntegration.yaml" };
+		String[] expectedPipelinefile = { "testWebhook.yaml" };
 
 		// VERIFY
 		AssertionsHelper.assertFolderContainsFiles(getTempFolderProjectPath(), expectedPipelinefile);
-	}
-
-	@Test
-	void testExportMissingCi() {
-		// GIVEN
-		List<CustomIntegrationVersion> cis = new ArrayList<>();
-		CustomIntegrationVersion ci = new CustomIntegrationVersion();
-		cis.add(ci);
-		ci.setName("testCustomIntegration");
-		ci.setStatus("DRAFT");
-
-		List<String> names = Arrays.asList(new String[] { "notMachingCi" });
-		when(restClient.getCustomIntegrations()).thenReturn(cis);
-		when(vraNgPackageDescriptor.getCustomIntegration()).thenReturn(names);
-
-		// TEST
-		store.exportContent();
-
-		String[] expectedPipelinefile = {};
-
-		// VERIFY
-		AssertionsHelper.assertFolderContainsFiles(getTempFolderProjectPath(), expectedPipelinefile);
-
 	}
 
 	@Test
 	void testImportContentIm() {
+		// GIVEN
+		List<JsonObject> pipelines = new ArrayList<>();
+		JsonObject pipeline = new JsonObject();
+		pipeline.addProperty("name", "newWebhook");
+		pipeline.addProperty("state", "ENABLED");
+		// pipelines.add(pipeline);
+		createTempFile("newWebhook", pipeline);
 
-		List<CustomIntegrationVersion> storageNew = new ArrayList<>();
-		CustomIntegrationVersion ciNew = new CustomIntegrationVersion();
-		storageNew.add(ciNew);
-		ciNew.setName("testCustomIntegrationNew");
-		ciNew.setStatus("DRAFT");
-		createTempFile("testCustomIntegationNew", storageNew);
+		pipeline = new JsonObject();
+		pipeline.addProperty("name", "existingWebhook");
+		pipeline.addProperty("state", "RELEASED");
+		pipeline.addProperty("id", "xxxx-yyyyy-xxxxx");
+		pipelines.add(pipeline);
+		createTempFile("existingWebhook", pipeline);
 
-		List<CustomIntegrationVersion> cis = new ArrayList<>();
-		List<CustomIntegrationVersion> storage = new ArrayList<>();
-		CustomIntegrationVersion ci = new CustomIntegrationVersion();
-		cis.add(ci);
-		storage.add(ci);
-		ci.setName("testCustomIntegration");
-		ci.setStatus("DRAFT");
-
-		List<CustomIntegrationVersion> versions = new ArrayList<>();
-		CustomIntegrationVersion version = new CustomIntegrationVersion();
-		versions.add(version);
-		storage.add(version);
-		version.setName("testCustomIntegration");
-		version.setVersion("1");
-		version.setStatus("RELEASED");
-		version = new CustomIntegrationVersion();
-		storage.add(version);
-		version.setName("testCustomIntegration");
-		version.setVersion("2");
-		version.setStatus("RELEASED");
-		createTempFile("testCustomIntegration", storage);
-
-		List<String> names = Arrays.asList(new String[] { "testCustomIntegration" });
-
-		when(restClient.getCustomIntegrations()).thenReturn(cis);
-		when(restClient.getCustomIntegrationVersions(any())).thenReturn(versions);
-		when(vraNgPackageDescriptor.getCustomIntegration()).thenReturn(names);
+		when(restClient.getProjectGitWebhooks()).thenReturn(pipelines);
 		when(restClient.getProjectName()).thenReturn("myProject");
 		// TEST
 		store.importContent(tempFolder.getRoot());
 
 		// VERIFY
-		verify(restClient, times(1)).createCustomIntegration(any());
-		verify(restClient, times(1)).updateCustomIntegration(any());
-		verify(restClient, times(1)).createCustomIntegrationVersion(any(), any());
+		verify(restClient, times(1)).createGitWebhook(any());
+		verify(restClient, times(1)).updateGitWebhook(any(), any());
+
 	}
 
 	/**
@@ -208,13 +156,14 @@ public class CsCustomIntegrationStoreTest {
 		return files[0];
 	}
 
-	private void createTempFile(String name, List<CustomIntegrationVersion> obj) {
-		File file = Paths.get(tempFolder.getRoot().getPath(), "custom-integrations", name + ".yaml").toFile();
+	private void createTempFile(String name, JsonObject obj) {
+		File file = Paths.get(tempFolder.getRoot().getPath(), "git-webhooks", name + ".yaml").toFile();
 		file.getParentFile().mkdirs();
 		try {
+			JsonNode jsonNodeTree = new ObjectMapper().readTree(obj.toString());
 			YAMLMapper yamlMapper = new YAMLMapper();
 			yamlMapper.setSerializationInclusion(Include.NON_NULL);
-			String yamlString = yamlMapper.writeValueAsString(obj);
+			String yamlString = yamlMapper.writeValueAsString(jsonNodeTree);
 			// StringWriter writer = new StringWriter();
 			// writer.write(triggerYaml);
 			Files.write(Paths.get(file.getPath()), yamlString.getBytes(),
